# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/leak-test": {
            "get": {
                "description": "Check the status of the leaks",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "leaks"
                ],
                "summary": "Check the status of the leaks",
                "responses": {
                    "200": {
                        "description": "Leak status",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/safe-test": {
            "get": {
                "description": "Start a safe goroutine that does not leak",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "leaks"
                ],
                "summary": "Start a safe goroutine",
                "responses": {
                    "200": {
                        "description": "Safe goroutine started",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/start-leak/{type}": {
            "get": {
                "description": "Start a memory leak of a specific type",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "leaks"
                ],
                "summary": "Start a memory leak",
                "parameters": [
                    {
                        "enum": [
                            "slice",
                            "channel",
                            "mutex"
                        ],
                        "type": "string",
                        "description": "Leak type",
                        "name": "type",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Leak started",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    }
}